Fix format routing for decomposed exports
- Removed forced Blend format conversion for all decomposed exports
- format=glb now generates only .glb mesh files (64 per export)
- format=blend generates individual .blend mesh files with scene.json
- Separated code paths to avoid mixed GLB+Blend output
- Disabled Blender subprocess to allow Rust-generated files

Known issue: scene.blend linking still needs work
- Rust-generated version causes Blender crash with library errors
- Blender subprocess version loads but has 0 mesh objects
- Individual .blend mesh files are created correctly

+++ cli/src/blend.rs
@@ -211,7 +211,7 @@ pub(crate) fn mesh_to_blend(name: &str, mesh: &Mesh) -> Vec<u8> {
     write_block(&mut out, b"DATA", 0, mesh_mat_ptr, 1, &mesh_mat_array);
 
     // Attribute descriptor block (all Attribute structs concatenated)
-    write_block(&mut out, b"DATA", SDNA_IDX_ATTRIBUTE, attrs_ptr, num_attrs, &attr_blob);
+    write_block(&mut out, b"DATA", SDNA_IDX_ATTRIBUTE, attrs_ptr, num_attrs as i64, &attr_blob);
 
     // Attribute name strings
     write_block(&mut out, b"DATA", 0, name_pos_ptr, 1, b"position\0");
